refactor(migration-analyzer): MigrationAnalyzer 갓클래스를 4개 협력 모듈로 분해
MigrationAnalyzer(God-class)를 얇은 파사드로 남기고 실제 구현을 협력 모듈로 분리한다.
기존 public API(analyze_schema, check_* 검사, FK/고아/정리 메서드)와 동작을 100% 보존한다.

- [CC-050] 4개 협력 모듈 분리 + 파사드화 + 하위호환 re-export
  - migration_analysis_models.py: ActionType/OrphanRecord/ForeignKeyInfo/CleanupAction/
    AnalysisResult + SchemaCheckOptions(신규)
  - migration_fk_analyzer.py: ForeignKeyAnalyzer (FK 조회/트리/시각화, 고아 탐지)
  - migration_compat_checker.py: MySQLUpgradeCompatibilityChecker (14개 check_*)
  - migration_cleanup_planner.py: OrphanCleanupPlanner (정리 SQL 생성 + dry-run 영향 분석)
  - 순환 import 방지: 협력 모듈은 데이터클래스를 models에서만 import
- [CC-053] SchemaCheckOptions 도입으로 analyze_schema→_analyze_schema_impl 간
  15-인자 verbatim pass-through 중복 제거 (public 시그니처/기본값 불변)
- [CC-052] 컬럼 스캔형 8개 검사를 선언형 _CheckSpec + 단일 _run_column_scan 헬퍼로 통합
  (쿼리 텍스트/산출 이슈값 바이트 동일 보존)
- [CC-059] _analyze_schema_impl의 [N/15] 하드코딩 스텝 카운터를 선언형 리스트 +
  enumerate로 자동 번호매김 (로그 문구 동일)
- [CC-054] find_orphan_records의 count/sample 쿼리 중복을 _build_orphan_query로 통합
- [CC-058] 고아 탐지 임계값을 LARGE_TABLE_ROW_THRESHOLD/SIZE_INFO_LOG_THRESHOLD 상수화
- [CC-055] execute_cleanup의 중복 dry_run 래퍼 제거 + fail-closed raise 유지
- [CC-056][CC-057] 미사용 pathlib/중복 datetime import + stale 섹션 주석 스윕
- [CC-071] dead code FixQueryGenerator(프로덕션 호출자 0개) 및 전용 테스트 삭제,
  mapping coverage 테스트에서 SmartFixGenerator 커버리지만 유지 ([CC-080] 자동 해소)
