+++ tests/test_migration_mapping_coverage.py
@@ -1,83 +1,19 @@
 """
 마이그레이션 종단간 매핑 커버리지 테스트
 
-IssueType → FixQueryGenerator → SmartFixGenerator 간 종단간 연결을 검증합니다.
-AutoRecommendationEngine(Python 자동 추천 엔진)은 Rust DB Core로 이관되어
-삭제되었으므로, 이 테스트는 더 이상 Python 추천 선택을 검증하지 않는다.
+IssueType → SmartFixGenerator 간 종단간 연결을 검증합니다.
+FixQueryGenerator(dead code)와 AutoRecommendationEngine(Python 자동 추천 엔진)은
+각각 삭제/Rust DB Core 이관되었으므로, 이 테스트는 SmartFixGenerator 커버리지만 검증한다.
 """
 import pytest
 
 from src.core.migration_constants import IssueType, DOC_LINKS, CompatibilityIssue
-from src.core.migration_fix_generator import FixQueryGenerator
 from src.core.migration_fix_wizard import (
     FixStrategy,
     SmartFixGenerator,
     create_wizard_steps,
 )
 from tests.conftest import FakeMySQLConnector, make_issue
-
-
-# ============================================================
-# FixQueryGenerator 매핑 커버리지
-# ============================================================
-class TestFixQueryGeneratorCoverage:
-    """FixQueryGenerator가 지원하는 IssueType 매핑 검증"""
-
-    # FixQueryGenerator.generate() 내부 handlers 키 목록
-    EXPECTED_FIX_GENERATOR_TYPES = {
-        IssueType.AUTH_PLUGIN_ISSUE,
-        IssueType.CHARSET_ISSUE,
-        IssueType.ZEROFILL_USAGE,
-        IssueType.FLOAT_PRECISION,
-        IssueType.INVALID_DATE,
-        IssueType.YEAR2_TYPE,
-        IssueType.DEPRECATED_ENGINE,
-        IssueType.ENUM_EMPTY_VALUE,
-        IssueType.INDEX_TOO_LARGE,
-        IssueType.FK_NAME_LENGTH,
-        IssueType.RESERVED_KEYWORD,
-        IssueType.INT_DISPLAY_WIDTH,
-        IssueType.LATIN1_CHARSET,
-        IssueType.FK_NON_UNIQUE_REF,
-        IssueType.SUPER_PRIVILEGE,
-        IssueType.REMOVED_SYS_VAR,
-        IssueType.GROUPBY_ASC_DESC,
-        IssueType.SQL_CALC_FOUND_ROWS_USAGE,
-        IssueType.PARTITION_ISSUE,
-        IssueType.TIMESTAMP_RANGE,
-        IssueType.BLOB_TEXT_DEFAULT,
-        IssueType.DEPRECATED_FUNCTION,
-        IssueType.SQL_MODE_ISSUE,
-        IssueType.FTS_TABLE_PREFIX,
-        IssueType.FK_REF_NOT_FOUND,
-    }
-
-    def test_all_expected_types_have_generator(self):
-        """FixQueryGenerator에 등록된 모든 타입이 실제 핸들러를 가지는지"""
-        gen = FixQueryGenerator()
-        for issue_type in self.EXPECTED_FIX_GENERATOR_TYPES:
-            issue = make_issue(
-                issue_type,
-                location="test_db.table.col",
-                table_name="table",
-                column_name="col",
-            )
-            result = gen.generate(issue)
-            # generate()는 issue를 반환 (fix_query가 추가될 수 있음)
-            assert isinstance(result, CompatibilityIssue)
-
-    @pytest.mark.parametrize("issue_type", list(IssueType))
-    def test_generate_never_raises(self, issue_type):
-        """모든 IssueType에 대해 generate()가 예외를 발생시키지 않음"""
-        gen = FixQueryGenerator()
-        issue = make_issue(
-            issue_type,
-            location="test_db.table.col",
-            table_name="table",
-            column_name="col",
-        )
-        result = gen.generate(issue)
-        assert isinstance(result, CompatibilityIssue)
 
 
 # ============================================================
@@ -172,15 +108,15 @@ class TestDocLinksCoverage:
 # 종단간 Flow 테스트
 # ============================================================
 class TestEndToEndFlow:
-    """IssueType → FixQueryGenerator → SmartFixGenerator → create_wizard_steps 흐름
+    """IssueType → SmartFixGenerator → create_wizard_steps 흐름
 
-    AutoRecommendationEngine(Python 자동 추천 선택)은 Rust DB Core로
-    이관되어 삭제되었으므로, 여기서는 create_wizard_steps까지만 검증하고
+    FixQueryGenerator(dead code)와 AutoRecommendationEngine(Python 자동 추천 선택)은
+    각각 삭제/Rust DB Core 이관되었으므로, 여기서는 create_wizard_steps까지만 검증하고
     Python 추천 선택 결과는 assert하지 않는다.
     """
 
     def test_invalid_date_full_flow(self):
-        """INVALID_DATE: 발견 → fix_query → wizard options"""
+        """INVALID_DATE: 발견 → wizard options"""
         conn = FakeMySQLConnector()
         conn.query_results = {
             'IS_NULLABLE': [{'IS_NULLABLE': 'YES'}],
@@ -194,19 +130,14 @@ class TestEndToEndFlow:
             column_name="created_at",
         )
 
-        # Step 1: FixQueryGenerator
-        gen = FixQueryGenerator()
-        gen.generate(issue)
-        assert issue.fix_query is not None
-
-        # Step 2: SmartFixGenerator
+        # SmartFixGenerator
         wizard_gen = SmartFixGenerator(conn, "test_db")
         options = wizard_gen.get_fix_options(issue)
         assert len(options) >= 2
         assert any(o.strategy == FixStrategy.DATE_TO_NULL for o in options)
 
     def test_charset_full_flow(self):
-        """CHARSET_ISSUE: 발견 → fix_query → wizard options"""
+        """CHARSET_ISSUE: 발견 → wizard options"""
         conn = FakeMySQLConnector()
         conn.query_results = {
             'KEY_COLUMN_USAGE': [],
@@ -218,11 +149,7 @@ class TestEndToEndFlow:
             table_name="users",
         )
 
-        # Step 1: FixQueryGenerator
-        gen = FixQueryGenerator()
-        gen.generate(issue)
-
-        # Step 2: SmartFixGenerator
+        # SmartFixGenerator
         wizard_gen = SmartFixGenerator(conn, "test_db")
         options = wizard_gen.get_fix_options(issue)
         assert any(o.strategy == FixStrategy.COLLATION_SINGLE for o in options)
@@ -239,16 +166,13 @@ class TestEndToEndFlow:
             column_name="stock",
         )
 
-        gen = FixQueryGenerator()
-        gen.generate(issue)
-
         wizard_gen = SmartFixGenerator(conn, "test_db")
         options = wizard_gen.get_fix_options(issue)
         assert len(options) >= 1
         assert options[-1].strategy == FixStrategy.SKIP
 
     def test_deprecated_engine_full_flow(self):
-        """DEPRECATED_ENGINE: 발견 → fix_query → wizard options"""
+        """DEPRECATED_ENGINE: 발견 → wizard options"""
         conn = FakeMySQLConnector()
         conn.query_results = {'KEY_COLUMN_USAGE': []}
 
@@ -257,9 +181,6 @@ class TestEndToEndFlow:
             location="test_db.logs",
             table_name="logs",
         )
-
-        gen = FixQueryGenerator()
-        gen.generate(issue)
 
         wizard_gen = SmartFixGenerator(conn, "test_db")
         options = wizard_gen.get_fix_options(issue)
